requirements.txt
discord-ext-bot==1.0.1
discord.py==2.3.2
flask